fix: create utility to group addresses
