+++ packages/pn-personafisica-webapp/src/navigation/OnboardingGuard.tsx
@@ -7,13 +7,12 @@ import { LoadingPage, NotificationStatus } from '@pagopa-pn/pn-commons';
 import { OnboardingSource } from '../models/Onboarding';
 import { setIsFreshLogin } from '../redux/auth/reducers';
 import { getDigitalAddresses } from '../redux/contact/actions';
-import { contactsSelectors } from '../redux/contact/reducers';
 import { getReceivedNotifications } from '../redux/dashboard/actions';
 import { setFirstSearch } from '../redux/dashboard/reducers';
 import { useAppDispatch, useAppSelector } from '../redux/hooks';
 import { setOnboardingSource } from '../redux/sidemenu/reducers';
 import { getConfiguration } from '../services/configuration.service';
-import { hasRequiredContacts } from '../utility/contacts.utility';
+import { groupDigitalAddresses, hasRequiredContacts } from '../utility/contacts.utility';
 import * as routes from './routes.const';
 
 const hasNotificationsToRead = (notifications: Array<Notification>): boolean => {
@@ -32,21 +31,24 @@ const OnboardingGuard: React.FC = () => {
   const location = useLocation();
 
   const { IS_ONBOARDING_ENABLED } = getConfiguration();
-  const addresses = useAppSelector(contactsSelectors.selectAddresses);
   const isFreshLogin = useAppSelector((state) => state.userState.isFreshLogin);
 
   const [isLoading, setIsLoading] = useState(true);
 
   const fetchOnboardingData = async () => {
-    await dispatch(getDigitalAddresses()).unwrap();
+    const addresses = await dispatch(getDigitalAddresses()).unwrap();
     const notifications = await dispatch(getReceivedNotifications({ size: 10 })).unwrap();
     dispatch(setFirstSearch(true));
-    return notifications.resultsPage;
+
+    return {
+      notifications: notifications.resultsPage,
+      addresses: groupDigitalAddresses(addresses),
+    };
   };
 
   useEffect(() => {
     fetchOnboardingData()
-      .then((notifications) => {
+      .then(({ notifications, addresses }) => {
         const shouldRedirectToOnboarding =
           location.pathname === '/' &&
           isFreshLogin &&
